// Frey/app/clients/proto/frey.pb.ts
export interface AccessGetEventAdminsPayload {
  eventId: number;
}

export interface AccessGetEventAdminsResponse {
  admins: protoAtoms.EventAdmin[];
}

export interface AccessGetEventRefereesPayload {
  eventId: number;
}

export interface AccessGetEventRefereesResponse {
  referees: protoAtoms.EventReferee[];
}
// ...
export interface AccessAddRuleForPersonPayload {
  ruleName: string;
  ruleValue: protoAtoms.RuleValue;
  ruleType: string;
  personId: number;
  eventId: number;
}

export interface AccessAddRuleForPersonResponse {
  ruleId: number;
}
// ...
export interface AccessDeleteRuleForPersonPayload {
  ruleId: number;
}

export interface AccessDeleteRuleForGroupPayload {
  ruleId: number;
}

export interface AccessClearAccessCachePayload {
  personId: number;
  eventId: number;
}
// ...
export const AccessGetEventAdminsPayload = {
  /**
   * Serializes AccessGetEventAdminsPayload to protobuf.
   */
  encode: function (msg: PartialDeep<AccessGetEventAdminsPayload>): Uint8Array {
    return AccessGetEventAdminsPayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AccessGetEventAdminsPayload from protobuf.
   */
  decode: function (bytes: ByteSource): AccessGetEventAdminsPayload {
    return AccessGetEventAdminsPayload._readMessage(
      AccessGetEventAdminsPayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AccessGetEventAdminsPayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AccessGetEventAdminsPayload>,
  ): AccessGetEventAdminsPayload {
    return {
      eventId: 0,
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AccessGetEventAdminsPayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.eventId) {
      writer.writeInt32(1, msg.eventId);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AccessGetEventAdminsPayload,
    reader: protoscript.BinaryReader,
  ): AccessGetEventAdminsPayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.eventId = reader.readInt32();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const AccessGetEventAdminsResponse = {
  /**
   * Serializes AccessGetEventAdminsResponse to protobuf.
   */
  encode: function (
    msg: PartialDeep<AccessGetEventAdminsResponse>,
  ): Uint8Array {
    return AccessGetEventAdminsResponse._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AccessGetEventAdminsResponse from protobuf.
   */
  decode: function (bytes: ByteSource): AccessGetEventAdminsResponse {
    return AccessGetEventAdminsResponse._readMessage(
      AccessGetEventAdminsResponse.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AccessGetEventAdminsResponse with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AccessGetEventAdminsResponse>,
  ): AccessGetEventAdminsResponse {
    return {
      admins: [],
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AccessGetEventAdminsResponse>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.admins?.length) {
      writer.writeRepeatedMessage(
        1,
        msg.admins as any,
        protoAtoms.EventAdmin._writeMessage,
      );
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AccessGetEventAdminsResponse,
    reader: protoscript.BinaryReader,
  ): AccessGetEventAdminsResponse {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          const m = protoAtoms.EventAdmin.initialize();
          reader.readMessage(m, protoAtoms.EventAdmin._readMessage);
          msg.admins.push(m);
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};
// ...
export const AccessAddRuleForPersonPayload = {
  /**
   * Serializes AccessAddRuleForPersonPayload to protobuf.
   */
  encode: function (
    msg: PartialDeep<AccessAddRuleForPersonPayload>,
  ): Uint8Array {
    return AccessAddRuleForPersonPayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AccessAddRuleForPersonPayload from protobuf.
   */
  decode: function (bytes: ByteSource): AccessAddRuleForPersonPayload {
    return AccessAddRuleForPersonPayload._readMessage(
      AccessAddRuleForPersonPayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AccessAddRuleForPersonPayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AccessAddRuleForPersonPayload>,
  ): AccessAddRuleForPersonPayload {
    return {
      ruleName: "",
      ruleValue: protoAtoms.RuleValue.initialize(),
      ruleType: "",
      personId: 0,
      eventId: 0,
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AccessAddRuleForPersonPayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.ruleName) {
      writer.writeString(1, msg.ruleName);
    }
    if (msg.ruleValue) {
      writer.writeMessage(2, msg.ruleValue, protoAtoms.RuleValue._writeMessage);
    }
    if (msg.ruleType) {
      writer.writeString(3, msg.ruleType);
    }
    if (msg.personId) {
      writer.writeInt32(4, msg.personId);
    }
    if (msg.eventId) {
      writer.writeInt32(5, msg.eventId);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AccessAddRuleForPersonPayload,
    reader: protoscript.BinaryReader,
  ): AccessAddRuleForPersonPayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.ruleName = reader.readString();
          break;
        }
        case 2: {
          reader.readMessage(msg.ruleValue, protoAtoms.RuleValue._readMessage);
          break;
        }
        case 3: {
          msg.ruleType = reader.readString();
          break;
        }
        case 4: {
          msg.personId = reader.readInt32();
          break;
        }
        case 5: {
          msg.eventId = reader.readInt32();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};

export const AccessAddRuleForPersonResponse = {
  /**
   * Serializes AccessAddRuleForPersonResponse to protobuf.
   */
  encode: function (
    msg: PartialDeep<AccessAddRuleForPersonResponse>,
  ): Uint8Array {
    return AccessAddRuleForPersonResponse._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AccessAddRuleForPersonResponse from protobuf.
   */
  decode: function (bytes: ByteSource): AccessAddRuleForPersonResponse {
    return AccessAddRuleForPersonResponse._readMessage(
      AccessAddRuleForPersonResponse.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AccessAddRuleForPersonResponse with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AccessAddRuleForPersonResponse>,
  ): AccessAddRuleForPersonResponse {
    return {
      ruleId: 0,
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AccessAddRuleForPersonResponse>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.ruleId) {
      writer.writeInt32(1, msg.ruleId);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AccessAddRuleForPersonResponse,
    reader: protoscript.BinaryReader,
  ): AccessAddRuleForPersonResponse {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.ruleId = reader.readInt32();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};
// ...
export const AccessDeleteRuleForPersonPayload = {
  /**
   * Serializes AccessDeleteRuleForPersonPayload to protobuf.
   */
  encode: function (
    msg: PartialDeep<AccessDeleteRuleForPersonPayload>,
  ): Uint8Array {
    return AccessDeleteRuleForPersonPayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AccessDeleteRuleForPersonPayload from protobuf.
   */
  decode: function (bytes: ByteSource): AccessDeleteRuleForPersonPayload {
    return AccessDeleteRuleForPersonPayload._readMessage(
      AccessDeleteRuleForPersonPayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AccessDeleteRuleForPersonPayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AccessDeleteRuleForPersonPayload>,
  ): AccessDeleteRuleForPersonPayload {
    return {
      ruleId: 0,
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AccessDeleteRuleForPersonPayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.ruleId) {
      writer.writeInt32(1, msg.ruleId);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AccessDeleteRuleForPersonPayload,
    reader: protoscript.BinaryReader,
  ): AccessDeleteRuleForPersonPayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.ruleId = reader.readInt32();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};
// ...
export const AccessClearAccessCachePayload = {
  /**
   * Serializes AccessClearAccessCachePayload to protobuf.
   */
  encode: function (
    msg: PartialDeep<AccessClearAccessCachePayload>,
  ): Uint8Array {
    return AccessClearAccessCachePayload._writeMessage(
      msg,
      new protoscript.BinaryWriter(),
    ).getResultBuffer();
  },

  /**
   * Deserializes AccessClearAccessCachePayload from protobuf.
   */
  decode: function (bytes: ByteSource): AccessClearAccessCachePayload {
    return AccessClearAccessCachePayload._readMessage(
      AccessClearAccessCachePayload.initialize(),
      new protoscript.BinaryReader(bytes),
    );
  },

  /**
   * Initializes AccessClearAccessCachePayload with all fields set to their default value.
   */
  initialize: function (
    msg?: Partial<AccessClearAccessCachePayload>,
  ): AccessClearAccessCachePayload {
    return {
      personId: 0,
      eventId: 0,
      ...msg,
    };
  },

  /**
   * @private
   */
  _writeMessage: function (
    msg: PartialDeep<AccessClearAccessCachePayload>,
    writer: protoscript.BinaryWriter,
  ): protoscript.BinaryWriter {
    if (msg.personId) {
      writer.writeInt32(1, msg.personId);
    }
    if (msg.eventId) {
      writer.writeInt32(2, msg.eventId);
    }
    return writer;
  },

  /**
   * @private
   */
  _readMessage: function (
    msg: AccessClearAccessCachePayload,
    reader: protoscript.BinaryReader,
  ): AccessClearAccessCachePayload {
    while (reader.nextField()) {
      const field = reader.getFieldNumber();
      switch (field) {
        case 1: {
          msg.personId = reader.readInt32();
          break;
        }
        case 2: {
          msg.eventId = reader.readInt32();
          break;
        }
        default: {
          reader.skipField();
          break;
        }
      }
    }
    return msg;
  },
};